test: keep the cpuMultiplier tests with the other Cpu6502 tests
Also coerce the scaled key hold to an integer: cpuMultiplier is a float, and
the hold feeds runFor() and so execute()'s target cycle count.

Addresses review feedback on #725.
